+++ src/owi.mli
@@ -29,7 +29,6 @@ module Result : sig
     | `Malformed_utf8_encoding of string
     | `Memory_size_too_large
     | `Msg of string
-    | `Multiple_memories
     | `Multiple_start_sections
     | `No_error
     | `Parse_fail of string
@@ -314,21 +313,21 @@ module Text : sig
     | Table_init of indice * indice
     | Elem_drop of indice
     (* Memory instructions *)
-    | I_load of nn * memarg
-    | F_load of nn * memarg
-    | I_store of nn * memarg
-    | F_store of nn * memarg
-    | I_load8 of nn * sx * memarg
-    | I_load16 of nn * sx * memarg
-    | I64_load32 of sx * memarg
-    | I_store8 of nn * memarg
-    | I_store16 of nn * memarg
-    | I64_store32 of memarg
-    | Memory_size
-    | Memory_grow
-    | Memory_fill
-    | Memory_copy
-    | Memory_init of indice
+    | I_load of indice * nn * memarg
+    | F_load of indice * nn * memarg
+    | I_store of indice * nn * memarg
+    | F_store of indice * nn * memarg
+    | I_load8 of indice * nn * sx * memarg
+    | I_load16 of indice * nn * sx * memarg
+    | I64_load32 of indice * sx * memarg
+    | I_store8 of indice * nn * memarg
+    | I_store16 of indice * nn * memarg
+    | I64_store32 of indice * memarg
+    | Memory_size of indice
+    | Memory_grow of indice
+    | Memory_fill of indice
+    | Memory_copy of indice * indice
+    | Memory_init of indice * indice
     | Data_drop of indice
     (* Control instructions *)
     | Nop
@@ -555,21 +554,21 @@ module Binary : sig
     | Table_init of indice * indice
     | Elem_drop of indice
     (* Memory instructions *)
-    | I_load of Text.nn * Text.memarg
-    | F_load of Text.nn * Text.memarg
-    | I_store of Text.nn * Text.memarg
-    | F_store of Text.nn * Text.memarg
-    | I_load8 of Text.nn * Text.sx * Text.memarg
-    | I_load16 of Text.nn * Text.sx * Text.memarg
-    | I64_load32 of Text.sx * Text.memarg
-    | I_store8 of Text.nn * Text.memarg
-    | I_store16 of Text.nn * Text.memarg
-    | I64_store32 of Text.memarg
-    | Memory_size
-    | Memory_grow
-    | Memory_fill
-    | Memory_copy
-    | Memory_init of indice
+    | I_load of indice * Text.nn * Text.memarg
+    | F_load of indice * Text.nn * Text.memarg
+    | I_store of indice * Text.nn * Text.memarg
+    | F_store of indice * Text.nn * Text.memarg
+    | I_load8 of indice * Text.nn * Text.sx * Text.memarg
+    | I_load16 of indice * Text.nn * Text.sx * Text.memarg
+    | I64_load32 of indice * Text.sx * Text.memarg
+    | I_store8 of indice * Text.nn * Text.memarg
+    | I_store16 of indice * Text.nn * Text.memarg
+    | I64_store32 of indice * Text.memarg
+    | Memory_size of indice
+    | Memory_grow of indice
+    | Memory_fill of indice
+    | Memory_copy of indice * indice
+    | Memory_init of indice * indice
     | Data_drop of indice
     (* Control instructions *)
     | Nop
@@ -743,7 +742,7 @@ module Concrete_extern_func : sig
 
     val unit : (lr, unit, unit) t
 
-    val memory : (l, mem, Concrete_memory.t) t
+    val memory : int -> (l, mem, Concrete_memory.t) t
 
     val externref : 'a Type.Id.t -> (lr, elt, 'a) t
   end