+++ test/script/reference.t
@@ -1,9 +1,6 @@
   $ owi script --no-exhaustion reference/address.wast
   $ owi script --no-exhaustion reference/align.wast
   $ owi script --no-exhaustion reference/binary-leb128.wast
-  $ owi script --no-exhaustion reference/binary.wast
-  owi: [ERROR] expected END opcode expected but got (unexpected end of section or function)
-  [54]
   $ owi script --no-exhaustion reference/block.wast
   $ owi script --no-exhaustion reference/br_if.wast
   $ owi script --no-exhaustion reference/br_table.wast
@@ -15,7 +12,6 @@
   $ owi script --no-exhaustion reference/const.wast
   $ owi script --no-exhaustion reference/conversions.wast
   $ owi script --no-exhaustion reference/custom.wast
-  $ owi script --no-exhaustion reference/data.wast
   $ owi script --no-exhaustion reference/elem.wast
   $ owi script --no-exhaustion reference/endianness.wast
   $ owi script --no-exhaustion reference/exports.wast
@@ -38,41 +34,21 @@
   $ owi script --no-exhaustion reference/i32.wast
   $ owi script --no-exhaustion reference/i64.wast
   $ owi script --no-exhaustion reference/if.wast
-  $ owi script --no-exhaustion reference/imports.wast
-  13
-  14
-  42
-  13
-  13
-  13
-  13
-  24
-  25
-  53
-  24
-  24
-  24
-  24
-  13
   $ owi script --no-exhaustion reference/inline-module.wast
   $ owi script --no-exhaustion reference/int_exprs.wast
   $ owi script --no-exhaustion reference/int_literals.wast
   $ owi script --no-exhaustion reference/labels.wast
   $ owi script --no-exhaustion reference/left-to-right.wast
   $ owi script --no-exhaustion reference/linking.wast
-  $ owi script --no-exhaustion reference/load.wast
   $ owi script --no-exhaustion reference/local_get.wast
   $ owi script --no-exhaustion reference/local_set.wast
   $ owi script --no-exhaustion reference/local_tee.wast
   $ owi script --no-exhaustion reference/loop.wast
   $ owi script --no-exhaustion reference/memory_copy.wast
   $ owi script --no-exhaustion reference/memory_fill.wast
-  $ owi script --no-exhaustion reference/memory_grow.wast
   $ owi script --no-exhaustion reference/memory_init.wast
   $ owi script --no-exhaustion reference/memory_redundancy.wast
-  $ owi script --no-exhaustion reference/memory_size.wast
   $ owi script --no-exhaustion reference/memory_trap.wast
-  $ owi script --no-exhaustion reference/memory.wast
   $ owi script --no-exhaustion reference/names.wast
   42
   123
@@ -90,7 +66,6 @@
   $ owi script --no-exhaustion reference/start.wast
   1
   2
-  $ owi script --no-exhaustion reference/store.wast
   $ owi script --no-exhaustion reference/switch.wast
   $ owi script --no-exhaustion reference/table_copy.wast
   $ owi script --no-exhaustion reference/table_fill.wast
@@ -116,3 +91,66 @@
   $ owi script --no-exhaustion reference/proposals/annotations/annotations.wast
   owi: [ERROR] unknown operator "\"(@x \\\"\") \"unclosed"
   [23]
+  $ owi script --no-exhaustion reference/proposals/multi-memory/address0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/address1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/align0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/binary.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/binary0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/data_drop0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/data.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/data0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/data1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/exports0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/float_exprs0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/float_exprs1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/float_memory0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/imports.wast
+  13
+  14
+  42
+  13
+  13
+  13
+  13
+  24
+  25
+  53
+  24
+  24
+  24
+  24
+  13
+  $ owi script --no-exhaustion reference/proposals/multi-memory/imports0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/imports1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/imports2.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/imports3.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/imports4.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/linking0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/linking1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/linking2.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/linking3.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/load.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/load0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/load1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/load2.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_copy0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_copy1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_fill0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_grow.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_init0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_size.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_size0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_size1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_size2.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_size3.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_trap0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory_trap1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory-multi.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/memory.wast
+;  $ owi script --no-exhaustion reference/proposals/multi-memory/simd_memory-multi.wast
+; TODO: simd instructions are not yet fully supported
+  $ owi script --no-exhaustion reference/proposals/multi-memory/start0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/store.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/store0.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/store1.wast
+  $ owi script --no-exhaustion reference/proposals/multi-memory/traps0.wast